+++ db/cache.go
@@ -17,7 +17,7 @@ type userDataCache struct {
 
 	gatewayEndpoints     map[user.EndpointID]user.GatewayEndpoint
 	cacheRefreshInterval time.Duration
-	mu                   sync.RWMutex
+	gatewayEndpointsMu   sync.RWMutex
 
 	logger polylog.Logger
 }
@@ -28,12 +28,12 @@ func NewUserDataCache(driver DBDriver, cacheRefreshInterval time.Duration, logge
 
 		gatewayEndpoints:     make(map[user.EndpointID]user.GatewayEndpoint),
 		cacheRefreshInterval: cacheRefreshInterval,
-		mu:                   sync.RWMutex{},
+		gatewayEndpointsMu:   sync.RWMutex{},
 
 		logger: logger.With("component", "user_data_cache"),
 	}
 
-	if err := cache.setCache(context.Background()); err != nil {
+	if err := cache.updateCache(context.Background()); err != nil {
 		return nil, fmt.Errorf("failed to set cache: %w", err)
 	}
 
@@ -43,33 +43,33 @@ func NewUserDataCache(driver DBDriver, cacheRefreshInterval time.Duration, logge
 }
 
 func (c *userDataCache) GetGatewayEndpoint(ctx context.Context, endpointID user.EndpointID) (user.GatewayEndpoint, bool) {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
+	c.gatewayEndpointsMu.RLock()
+	defer c.gatewayEndpointsMu.RUnlock()
 
 	gatewayEndpoint, ok := c.gatewayEndpoints[endpointID]
 	return gatewayEndpoint, ok
 }
 
+// cacheRefreshHandler is intended to be run in a go routine.
 func (c *userDataCache) cacheRefreshHandler(ctx context.Context) {
 	for {
 		<-time.After(c.cacheRefreshInterval)
 
-		err := c.setCache(ctx)
-		if err != nil {
+		if err := c.updateCache(ctx); err != nil {
 			c.logger.Error().Err(err).Msg("failed to refresh cache")
 		}
 	}
 }
 
-func (c *userDataCache) setCache(ctx context.Context) error {
+func (c *userDataCache) updateCache(ctx context.Context) error {
 	gatewayEndpoints, err := c.db.GetGatewayEndpoints(ctx)
 	if err != nil {
 		return fmt.Errorf("failed to get gateway endpoints: %w", err)
 	}
 
-	c.mu.Lock()
+	c.gatewayEndpointsMu.Lock()
+	defer c.gatewayEndpointsMu.Unlock()
 	c.gatewayEndpoints = gatewayEndpoints
-	c.mu.Unlock()
 
 	return nil
 }
