+++ db/cache_test.go
@@ -101,14 +101,14 @@ func Test_cacheRefreshHandler(t *testing.T) {
 	}
 }
 
-func Test_setCache(t *testing.T) {
+func Test_updateCache(t *testing.T) {
 	tests := []struct {
 		name       string
 		mockReturn map[user.EndpointID]user.GatewayEndpoint
 		expected   map[user.EndpointID]user.GatewayEndpoint
 	}{
 		{
-			name:       "should set cache with gateway endpoints",
+			name:       "should update cache with gateway endpoints",
 			mockReturn: map[user.EndpointID]user.GatewayEndpoint{"endpoint_1": {EndpointID: "endpoint_1"}},
 			expected:   map[user.EndpointID]user.GatewayEndpoint{"endpoint_1": {EndpointID: "endpoint_1"}},
 		},
@@ -130,7 +130,7 @@ func Test_setCache(t *testing.T) {
 			cache, err := NewUserDataCache(mockDB, time.Minute, polyzero.NewLogger())
 			c.NoError(err)
 
-			err = cache.setCache(context.Background())
+			err = cache.updateCache(context.Background())
 			c.NoError(err)
 			c.Equal(test.expected, cache.gatewayEndpoints)
 		})
