+++ user/gateway_endpoint.go
@@ -1,6 +1,7 @@
 package user
 
-// A unique identifier for a user account's endpoint. It must be passed as the last path segment of the service request URL.
+// A unique identifier for a user account's GatewayEndpoint. It must be passed as the last path segment of
+// the service request URL. For example: POST https://api.path.xyz/v1/{gateway_endpoint_id}
 type EndpointID string
 
 // A unique identifier for a user's account, which identifies the account that owns the GatewayEndpoint.
@@ -11,16 +12,16 @@ type PlanType string
 
 // A GatewayEndpoint represents a user account's endpoint, which has two primary functions:
 // 1. Identifying which endpoint is being used to make a service request.
-// 2. Allowing configuration of endpoint-specific settings, such as API key authentication, etc.
+// 2. Allowing configuration of endpoint-specific settings, such as API key authorization, etc.
 //
 // A GatewayEndpoint is associated to a single UserAccount. A UserAccount can have multiple GatewayEndpoints.
-// Settings related to service requests, such as enforcing API key authentication, are configured per GatewayEndpoint.
+// Settings related to service requests, such as enforcing API key authorization, are configured per GatewayEndpoint.
 type GatewayEndpoint struct {
 	// The unique identifier for a user account's endpoint, which must be passed
 	// as the last path segment of the service request URL. eg `/v1/{endpoint_id}`
 	// This is used to identify the GatewayEndpoint when making a service request.
 	EndpointID EndpointID
-	// The authentication settings for the GatewayEndpoint.
+	// The authorization settings for the GatewayEndpoint.
 	Auth Auth
 	// The UserAccount that the GatewayEndpoint belongs to, including the PlanType.
 	UserAccount UserAccount
@@ -29,12 +30,12 @@ type GatewayEndpoint struct {
 	RateLimiting RateLimiting
 }
 
-// The authentication settings for a GatewayEndpoint.
+// The authorization settings for a GatewayEndpoint.
 type Auth struct {
 	// The API key for GatewayEndpoint. If APIKeyRequired is true, the API key
 	// must be passed as the `Authorization` HTTP header in service requests.
 	APIKey string
-	// Whether the GatewayEndpoint requires an API key for authentication.
+	// Whether the GatewayEndpoint requires an API key for authorization.
 	// If not set, the GatewayEndpoint does not require an API key to be passed
 	// as the `Authorization` header and all requests using the endpoint will be allowed.
 	APIKeyRequired bool
@@ -51,8 +52,10 @@ type UserAccount struct {
 // The rate limiting settings for a GatewayEndpoint.
 type RateLimiting struct {
 	// ThroughputLimit refers to rate limiting per-second (TPS).
+	// This is used to prevent DoS or DDoS attacks, as well as enforce pricing plan limits.
 	ThroughputLimit int
 	// CapacityLimit refers to rate limiting over longer periods, such as a day, week or month.
+	// This is to prevent abuse of the services provided, as well enforce pricing plan limits.
 	CapacityLimit int
 	// The period over which the CapacityLimit is enforced. One of `daily`, `weekly` or `monthly`.
 	CapacityLimitPeriod CapacityLimitPeriod
@@ -69,14 +72,20 @@ const (
 	CapacityLimitPeriodMonthly CapacityLimitPeriod = "monthly"
 )
 
+// GetAuth returns a the API key string and a boolean indicating
+// whether the GatewayEndpoint requires the API key for authorization.
 func (e *GatewayEndpoint) GetAuth() (string, bool) {
 	return e.Auth.APIKey, e.Auth.APIKeyRequired
 }
 
+// GetThroughputLimit returns the throughput limit (TPS) for the GatewayEndpoint,
+// which is the maximum number of requests per second that the GatewayEndpoint can handle.
 func (e *GatewayEndpoint) GetThroughputLimit() int {
 	return e.RateLimiting.ThroughputLimit
 }
 
+// GetCapacityLimit returns the capacity limit for the GatewayEndpoint and the period over
+// which it is enforced. eg. '100,000' & 'daily' = 100,000 requests per day.
 func (e *GatewayEndpoint) GetCapacityLimit() (int, CapacityLimitPeriod) {
 	return e.RateLimiting.CapacityLimit, e.RateLimiting.CapacityLimitPeriod
 }
